test: Rename test_autodoc.py to test_ext_autodoc.py
