+++ tests/test_ext_autodoc_events.py
@@ -11,7 +11,7 @@
 import pytest
 
 from sphinx.ext.autodoc import between, cut_lines
-from test_autodoc import do_autodoc
+from test_ext_autodoc import do_autodoc
 
 
 @pytest.mark.sphinx('html', testroot='ext-autodoc')
