+++ tests/test_ext_autodoc_private_members.py
@@ -10,7 +10,7 @@
 
 import pytest
 
-from test_autodoc import do_autodoc
+from test_ext_autodoc import do_autodoc
 
 
 @pytest.mark.sphinx('html', testroot='ext-autodoc')
